requirements.txt
behave==1.2.6
playwright==1.41.2
allure-behave==2.13.2
parse==1.20.1
parse-type==0.6.2
six==1.16.0
